requirements.txt
fastapi
uvicorn
google-api-python-client
google-auth
google-auth-oauthlib
sentence-transformers
faiss-cpu
numpy
pypdf
PyPDF2
python-multipart
openai